+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Version 3.6.6
+
+* Fixed Fluid/Ore Laser drill JEI recipe closes #1467
+* Fixed Fluid Laser Drill recipe requiring nether dimension and specific biomes
+
 # Version 3.6.5
 
 * Fixed Meat Feeder not showing in JEI